WIP: Regular Grammar string validation started, need to review some cases, and do a better version.

+++ PyFlap/grammars/views.py
@@ -34,7 +34,7 @@ def grammars_page(request):
         elif(form2.is_valid()): # SE FOR O FORM DE TESTAR A GRAMATICA
             context['grammartestform'] = form2
             context['grammarform'] = GrammarForm()
-            grammar.test_sentence(sentence=form2['testString'].value()) # Funcao que verifica a validade da sentença de acordo com a gramatica existente
+            grammar.test_sentence(sentence=form2['testString'].value(), index=None, currentNonTerminal='') # Funcao que verifica a validade da sentença de acordo com a gramatica existente
 
         else:
             print(f'\n\nNENHUM FORM VALIDO: {form1.is_valid()}\n\n')
@@ -51,9 +51,23 @@ def grammars_page(request):
     return render(request, 'grammars/index.html', context=context) # renderiza com as variáveis setadas
     
 
+def verify_rules_with_initial_symbol(current_rule):
+    rules = Rule.objects.all()
+    
+    grammar = Grammar.load()
+    for rule in rules:
+        if(rule != current_rule):
+            if(rule.left_side == grammar.initial):
+                return True
+    return False
 def delete_rule(request, id):
     rule = Rule.objects.get(pk=id)
-    rule.delete()
+    grammar = Grammar.load()
+    if(rule.left_side == grammar.initial): # Se tiver tentando remover uma regra com o simbolo inicial, garantimos que existe outra
+        if(verify_rules_with_initial_symbol(rule)):
+            rule.delete()
+    else:
+        rule.delete()
     return redirect('grammars_page')
 
 def delete_all_rules(request):
